// src/FE/src/api/generated/models/documentListResponse.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Workslip.Api | v1
 * OpenAPI spec version: 1.0.0
 */
import type { DocumentListItemResponse } from './documentListItemResponse';

export interface DocumentListResponse {
  items: DocumentListItemResponse[];
  /** @pattern ^-?(?:0|[1-9]\d*)$ */
  totalCount: number | string;
}